// src/app/enums/AllowanceTypeEnum.ts
export enum AllowanceTypeEnum {
    AllowanceIn = '5c60f693-bef5-e011-a485-80ee7300c694',
    AllowanceOut = '5c60f693-bef5-e011-a485-80ee7300c695',
    SignInallowance = '5c60f693-bef5-e011-a485-80ee7300c696',
    SignOutallowance = '5c60f693-bef5-e011-a485-80ee7300c697',
    SignInAndSignOutallowance = '5c60f693-bef5-e011-a485-80ee7300c698',
    AllowanceInAndOut = '5c60f693-bef5-e011-a485-80ee7300c699'
}